Change tag functions to work with txexpr-elements
Part of my frustrations with pollen came from using tag functions that
expected strings as arguments. Now that they work with txexpr-element?
as arguments, they can combine stuff freely without much trouble.

Also created a function `list-splice` to splice a list of elements into
a parent txexpr. The list isn't spliced until pollen processes the
txexpr (it just inserts a txexpr with the '@ symbol as the tag).

I also created a macro, let-splice, to use list-splice to place a bunch
of different things from a let body into a pollen document. There are
times when let bodies would be nice, like when you want to compute the
answers to an example directly in the document. These expressions can
quickly be extracted into a tag, too. The tag can just return a
let-splice expression.

I like using let because let allows me to do things like: create
temporary tag functions that are relevant just for a part of the
document. And do so right in the document. It allows me to describe
computations directly in the document, too.

Changes to be committed:
      modified:   latex-commands.rkt
      new file:   manual-traverse.rkt
      modified:   pollen.rkt
      modified:   template.tex.p
      modified:   utility.rkt